requirements.txt
importlib-metadata==8.0.0
importlib-resources==6.4.0
jaraco.collections==5.1.0
packaging==24.1
pip-chill==1.0.3
platformdirs==4.2.2
tomli==2.0.1
pyserial==3.5
ipykernel
matplotlib==3.9.2